# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "御臻商城API文档",
        "description": "",
        "contact": {
            "email": "xiaopei0206@icloud.com"
        },
        "version": "0.6.0"
    },
    "servers": [
        {
            "url": "http://zhenmall.local",
            "description": ""
        }
    ],
    "paths": {
...
        "/api/red-packets": {
            "get": {
                "tags": [
                    "RedPacket"
                ],
                "summary": "Get red packets 获取红包列表",
                "operationId": "759d5c946b49df06b8903e369369b77a",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/red-packets/{id}": {
            "put": {
                "tags": [
                    "RedPacket"
                ],
                "summary": "Open red packets 开红包",
                "operationId": "7d1d0e95e75649c8394c1858be7bfd51",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}